requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.22.0
httpx==0.24.1
pydantic==2.6.1
python-dotenv==1.0.0
pytest==7.4.2
httpx==0.24.1